issue_tracker/requirements.txt
amqp==5.2.0
anyio==4.3.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.4
django-simple-history==3.5.0
djangorestframework==3.15.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
kombu==5.3.7
packaging==24.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
redis==5.0.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
